Implement xs_write
Choice for the type of data accepted by Xs::write is not obvious.  The
Xenstore protocol cannot handle arbitrary data, and a NUL byte in the
data to be written will be taken as string terminator by the C API.

This commit uses a character slice for the written data, which looks
more idiomatic than [u8].  Both have the ability to embed NUL
characters, in which case any character following a NUL won't make it
to the Xenstore.

Maybe we could use instead of custom type that would ensure there is no
NUL in input data.  Leaving this for a possible future evolution.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>